+++ src/app/[locale]/layout.tsx
@@ -9,9 +9,9 @@ import { getMetadata, getViewport } from "@/lib/seo";
 import { getLocaleDirection } from "@/lib/locale/utils";
 import type { AppLocale } from "~/i18n.config";
 import { Toaster } from "@/components/ui/toaster";
-import Script from "next/script";
 import { env } from "@/env";
 import DemoModalProvider from "../_components/video-modal/provider";
+import Analytics from "./analytics";
 
 export async function generateMetadata() {
   return getMetadata({
@@ -50,48 +50,8 @@ export default function LocaleLayout({
           <DemoModalProvider />
         </Providers>
 
-        {process.env.NODE_ENV === "production" && <Analytics />}
+        {env.VERCEL_ENV === "production" && <Analytics />}
       </body>
     </html>
   );
 }
-
-const Analytics = () => {
-  return (
-    <>
-      {/* <GoogleAnalytics gaId={env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID} />
-      <GoogleTagManager gtmId={env.NEXT_PUBLIC_GOOGLE_TAG_MANAGER_ID} /> */}
-
-      <Script
-        src={`https://www.googletagmanager.com/gtag/js?id=${env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}`}
-        strategy="lazyOnload"
-        id="google-analytics-script-url"
-      />
-
-      <Script id="google-analytics-script-code" strategy="lazyOnload">
-        {`window.dataLayer = window.dataLayer || [];
-         function gtag(){dataLayer.push(arguments);}
-         gtag('js', new Date());
-         gtag('config', '${env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}', {
-            page_path: window.location.pathname,
-         });`}
-      </Script>
-
-      {env.NEXT_PUBLIC_ENABLE_CLARITY === "true" && (
-        <Script
-          id="clarity-script"
-          strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
-    (function(c,l,a,r,i,t,y){
-        c[a]=c[a]||function(){(c[a].q=c[a].q||[]).push(arguments)};
-        t=l.createElement(r);t.async=1;t.src="https://www.clarity.ms/tag/"+i;
-        y=l.getElementsByTagName(r)[0];y.parentNode.insertBefore(t,y);
-    })(window, document, "clarity", "script", "${env.NEXT_PUBLIC_CLARITY_PROJECT_ID}");
-    `,
-          }}
-        />
-      )}
-    </>
-  );
-};
